Removes test and moves adding functionality to a custom hook
