+++ src/components/AddList/AddList.scss
@@ -6,6 +6,7 @@
 	height: 41px;
 	margin: 0px 0px 0px 10px;
 	transition: height 85ms ease-out;
+	user-select: none;
 	width: 272px;
 	vertical-align: top;
 }
